+++ experimental/packages/otlp-grpc-exporter-base/src/types.ts
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 
-import type { ChannelCredentials, Metadata } from '@grpc/grpc-js';
+import type {
+  ChannelCredentials,
+  ChannelOptions,
+  Metadata,
+} from '@grpc/grpc-js';
 
 import {
   CompressionAlgorithm,
@@ -29,4 +33,5 @@ export interface OTLPGRPCExporterConfigNode extends OTLPExporterConfigBase {
   metadata?: Metadata;
   compression?: CompressionAlgorithm;
   userAgent?: string;
+  channelOptions?: ChannelOptions;
 }